requirements.txt
fastapi
uvicorn
pymongo
earthengine-api
pandas
scikit-learn
requests
python-dotenv
gunicorn
